+++ src/main/java/com/github/deeepamin/ciaid/refactor/dialogs/MoveJobDialog.java
@@ -1,8 +1,8 @@
 package com.github.deeepamin.ciaid.refactor.dialogs;
 
 import com.github.deeepamin.ciaid.CIAidBundle;
+import com.github.deeepamin.ciaid.utils.CIAidUtils;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.ui.panel.ComponentPanelBuilder;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.util.ui.FormBuilder;
 
@@ -21,17 +21,16 @@ public class MoveJobDialog extends BaseRefactorToFileDialog {
     return CIAidBundle.message("refactoring.move.job.dialog.title");
   }
 
-  @SuppressWarnings("deprecation")
   public void configureDialogContents() {
     super.configureDialogContents();
     var commentText = CIAidBundle.message("refactoring.move.job.dialog.comment.text");
-    moveJobCommentLabel = ComponentPanelBuilder.createCommentComponent(commentText, true);
+    moveJobCommentLabel = CIAidUtils.createCommentComponent(commentText, true);
   }
 
   @Override
   protected void customizeFormBuilder(FormBuilder formBuilder) {
     super.customizeFormBuilder(formBuilder);
-    formBuilder.setFormLeftIndent(75)
+    formBuilder.setFormLeftIndent(55)
             .addComponent(moveJobCommentLabel)
             .setFormLeftIndent(0);
   }